+++ packages/hooks/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.3.2+6
+
+ - Update a dependency to the latest release.
+
+## 0.3.2+5
+
+ - Update a dependency to the latest release.
+
+## 0.3.2+4
+
+ - Update a dependency to the latest release.
+
+## 0.3.2+3
+
+ - Update a dependency to the latest release.
+
 ## 0.3.2+2
 
  - Update a dependency to the latest release.