+++ packages/widgets/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.0.10+7
+
+ - Update a dependency to the latest release.
+
+## 0.0.10+6
+
+ - Update a dependency to the latest release.
+
+## 0.0.10+5
+
+ - Update a dependency to the latest release.
+
+## 0.0.10+4
+
+ - Update a dependency to the latest release.
+
 ## 0.0.10+3
 
  - Update a dependency to the latest release.